REFACTOR: Rewrite transport classes
The transport related classes have been reorganized into several classes that
echo does exactly what they are designed to.
It should be much simpler and clearer to support all kind of transport protocol
and device, including TAP/TUN device and pipes.

The `TransportInterface` use `std::function` as callbacks, which are not very
efficient but is the only choice if we want to use capture lambdas as callbacks.

The implementation of `TcpSocket` should move the unique pointer of `Buffer`
into the lambda callbacks, but Boost require the callback to be copy-able, so we
have to keep them with the `TcpSocket` and clean up them manually afterwards. As
such, there is probably no requirement for C++ 14 features as of now, but why
would anyone use ancient compiler.