FEAT: Disable boost handler type check
Boost check handler type so the type error will be more clear friendlier.
However, the check emits a error for lacking a copy constructor even if it is
never needed. Currently, if the handler is actually required to be copy
constructible, the error message will be much hander to understand.